Add a JSON schema for the LocationHistoryData type
Taken from the Android model [1].

I tested it with all of the data that had been uploaded from the Android
publisher example app to the S3 bucket (the data at the top level of the
bucket, not the ones in the "archive" directory).

The valid-android.json is file taken from the 1_2022-10-19_10/07/41 file
in that bucket (with the coordinates changed for privacy of whoever
recorded the data).

Closes #62.

[1] https://github.com/ably/ably-asset-tracking-android/blob/8016b0fb97be04332d0c4d39404933a944d444fc/publishing-sdk/src/main/java/com/ably/tracking/publisher/LocationHistoryModels.kt#L12-L18