Add rules for versioning of LocationHistoryData objects
The `pattern` regular expression comes from the Semantic Versioning FAQ
[1].

We decided (see conversation [2]) to move to using Semantic Versioning,
to be consistent with the approach used for ably/ably-common. This
breaks backwards compatibility with version 1 validators, which are
expecting a number in the `version` property. But we decided we’re okay
with this, since we think there’s nobody really using the location
history recording functionality in the wild at the moment, and if there
were, they’d be people who we could tell to just use a newer version of
the AAT library.

The version 2.0.0 schema is currently a pre-release, since I still want
to add a required `agents` property (see #63).

[1] https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string
[2] https://github.com/ably/ably-asset-tracking-common/pull/66#discussion_r1028145208